moved TaskModal
Seperated the files for easier debug. to TaskModal.js
